src/requirements.txt
numpy>=2.2.6
opencv-contrib-python>=4.12.0.88
ultralytics>=8.3.233
python-chess>=1.10.0
torch>=2.9.1
torchvision>=0.24.1
onnxruntime>=1.23.2
onnx>=1.15.0
onnxruntime>=1.18.0